增加分页功能

+++ property-common/src/main/java/cn/zut/common/dao/SelectCountByExample.java
@@ -7,5 +7,5 @@ import org.apache.ibatis.annotations.Param;
  */
 public interface SelectCountByExample<S> {
 
-    Long selectCountByExample(@Param("search") S search);
+    int selectCountByExample(@Param("search") S search);
 }
